+++ src/interface/ainterface.h
@@ -23,7 +23,7 @@ class AbstractInterface {
   const bool RunMenu(const std::vector<std::function<bool(void)>> &func, std::size_t menu);
   const std::size_t ShowMenu(const std::string &menu, const std::size_t items = 0U);
   
-  [[nodiscard]] const std::size_t CheckInputItem(const std::int64_t min, const std::int64_t max);
+  [[nodiscard]] const std::int64_t CheckInputItem(const std::int64_t min, const std::int64_t max);
   [[nodiscard]] std::pair<bool, std::filesystem::path> CheckInputPathFile();
   [[nodiscard]] const bool Exit();
